+++ apps/backend/src/app.module.ts
@@ -34,6 +34,7 @@ import { ItemsModule } from "./items/items.module";
 import { CommunityColorsModule } from "./community-colors/community-colors.module";
 import { PendingOwnershipModule } from "./pending-ownership/pending-ownership.module";
 import { DiscordModule } from "./discord/discord.module";
+import { QueueConsumerModule } from "./queue-consumer/queue-consumer.module";
 import { Request, Response } from "express";
 
 @Module({
@@ -214,6 +215,7 @@ import { Request, Response } from "express";
     CommunityColorsModule,
     PendingOwnershipModule,
     DiscordModule,
+    QueueConsumerModule,
   ],
 })
 export class AppModule {}